crawlchat-service/crawler-service/requirements.txt
# Common package (local)
-e ../common

# Core FastAPI and web framework
fastapi==0.103.2
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-core==2.10.1
email-validator==2.0.0
pydantic-settings==2.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Database
motor==3.3.0
pymongo==4.6.0

# AWS and cloud storage
boto3==1.34.0
botocore==1.34.0

# HTTP and networking
aiohttp==3.8.0
requests==2.31.0
urllib3==1.26.0
httpx==0.25.0

# Web scraping and parsing
beautifulsoup4==4.12.0
lxml==4.9.0
selenium==4.15.0
webdriver-manager==4.0.0

# Document processing - Basic text extraction only
PyPDF2==3.0.0
pdfminer.six==20221105
openpyxl==3.1.0
python-docx==1.1.0
python-pptx==0.6.21
html2text==2020.1.16
python-dateutil==2.8.2

# AWS Textract dependencies
Pillow==10.0.0

# AI and machine learning
openai==1.93.0

# Utilities
asyncio-throttle==1.0.2

# Monitoring and logging
prometheus-client==0.17.0

# Security and authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.5.0
PyJWT==2.8.0
cryptography==41.0.0

# AWS Lambda compatibility
mangum==0.19.0 